move a big chunk of stuff to arptypes.
only logs and dynamic things are left
