+++ arpeggio/src/scene.rs
@@ -1,12 +1,36 @@
 use std::collections::{HashMap, HashSet};
 
 use crate::{
-  grid::{make_world, query_world, TileSystemExt},
-  types::*,
+  game::GameExt, grid::{make_world, query_world, TileSystemExt}, types::*
 };
 
-impl Scene {
-  pub fn create(creation: SceneCreation) -> Scene {
+pub trait SceneExt {
+   fn create(creation: SceneCreation) -> Scene;
+   fn get_pos(&self, creature_id: CreatureID) -> Result<Point3, GameError>;
+   fn set_pos(&self, cid: CreatureID, pt: Point3) -> Result<Scene, GameError>;
+   fn add_volume_condition(
+    &self, condition_id: ConditionID, point: Point3, volume: Volume, condition: Condition,
+    duration: Duration,
+  ) -> Scene;
+
+  /// Figure out which volume conditions apply to the given creature.
+   fn creature_volume_conditions(
+    &self, game: &Game, creature: &Creature,
+  ) -> Result<Vec<(ConditionID, &VolumeCondition)>, GameError>;
+
+  /// Return a set of points of open terrain which  intersect a volume.
+  /// Largely used for previewing the area that will be affected by a volume-affecting ability.
+   fn open_terrain_in_volume(
+    &self, game: &Game, pt: Point3, volume: Volume,
+  ) -> Result<Terrain, GameError>;
+
+   fn creatures_in_volume(&self, ts: TileSystem, pt: Point3, volume: Volume) -> Vec<CreatureID>;
+
+   fn get_world(&self, game: &Game) -> Result<CollisionWorld, GameError>;
+}
+
+impl SceneExt for Scene {
+   fn create(creation: SceneCreation) -> Scene {
     Scene {
       id: SceneID::gen(),
       name: creation.name,
@@ -25,14 +49,14 @@ impl Scene {
       focused_creatures: vec![],
     }
   }
-  pub fn get_pos(&self, creature_id: CreatureID) -> Result<Point3, GameError> {
+   fn get_pos(&self, creature_id: CreatureID) -> Result<Point3, GameError> {
     self
       .creatures
       .get(&creature_id)
       .map(|x| x.0)
       .ok_or_else(|| GameError::CreatureNotFound(creature_id.to_string()))
   }
-  pub fn set_pos(&self, cid: CreatureID, pt: Point3) -> Result<Scene, GameError> {
+   fn set_pos(&self, cid: CreatureID, pt: Point3) -> Result<Scene, GameError> {
     let mut new = self.clone();
     {
       let data =
@@ -41,7 +65,7 @@ impl Scene {
     }
     Ok(new)
   }
-  pub fn add_volume_condition(
+   fn add_volume_condition(
     &self, condition_id: ConditionID, point: Point3, volume: Volume, condition: Condition,
     duration: Duration,
   ) -> Scene {
@@ -53,7 +77,7 @@ impl Scene {
   }
 
   /// Figure out which volume conditions apply to the given creature.
-  pub fn creature_volume_conditions(
+   fn creature_volume_conditions(
     &self, game: &Game, creature: &Creature,
   ) -> Result<Vec<(ConditionID, &VolumeCondition)>, GameError> {
     let condition_ids = query_world(&self.get_world(game)?, |cdata1, cdata2| {
@@ -82,19 +106,19 @@ impl Scene {
 
   /// Return a set of points of open terrain which  intersect a volume.
   /// Largely used for previewing the area that will be affected by a volume-affecting ability.
-  pub fn open_terrain_in_volume(
+   fn open_terrain_in_volume(
     &self, game: &Game, pt: Point3, volume: Volume,
   ) -> Result<Terrain, GameError> {
     let all_open = self.terrain.iter().map(|pt| (*pt, *pt)).collect();
     Ok(game.tile_system.items_within_volume(volume, pt, &all_open))
   }
 
-  pub fn creatures_in_volume(&self, ts: TileSystem, pt: Point3, volume: Volume) -> Vec<CreatureID> {
+   fn creatures_in_volume(&self, ts: TileSystem, pt: Point3, volume: Volume) -> Vec<CreatureID> {
     let creature_locations = self.creatures.iter().map(|(cid, &(pt, _))| (*cid, pt)).collect();
     ts.items_within_volume(volume, pt, &creature_locations)
   }
 
-  pub fn get_world(&self, game: &Game) -> Result<CollisionWorld, GameError> {
+   fn get_world(&self, game: &Game) -> Result<CollisionWorld, GameError> {
     let creatures = self.creatures.iter().filter_map(|(creature_id, &(pos, _))| {
       game.get_creature(*creature_id).map(|dc| (dc.creature, pos)).ok()
     });
@@ -106,8 +130,7 @@ impl Scene {
 #[cfg(test)]
 mod test {
   use crate::{
-    game::test::*,
-    types::{test::*, *},
+    game::test::*, scene::SceneExt, types::{test::*, *}
   };
 
   #[test]
