+++ worker/src/durablestorage.rs
@@ -8,7 +8,7 @@ use anyhow::anyhow;
 use tracing::{info, warn};
 use worker::{ListOptions, State};
 
-use arpeggio::types::{ChangedGame, Game, GameLog};
+use arpeggio::{game::GameExt, types::{ChangedGame, Game, GameLog}};
 use mtarp::types::{GameIndex, ImageType, InvitationID};
 
 use crate::anyhow_str;
